Create README.md
A successful recruiter with strong business acumen and interpersonal skills seeking assignments in Human Resource Management with a frontline organization. I enjoy building teams from scratch, bridging the gap between right job and individual and making the journey as seamless as possible for the hiring manager and the candidate, utilizing new and established methodologies and strategies to engage effectively and bring in the best talent.